pyproject.toml
[project]
name = "spreadsheet-mcp"
version = "0.1.0"
description = "MCP Server for Google Sheets - read, write, format cells, create charts, and more"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.0.0",
    "google-api-python-client>=2.150.0",
    "google-auth>=2.35.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
spreadsheet-mcp = "spreadsheet_mcp.server:main"

[project.optional-dependencies]
dev = [
    "ruff>=0.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/spreadsheet_mcp"]
